+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Add `getLocaleCode()` to StringTableLocale.
+- Add `indexResources()` to Package.
 
 ## [0.4.2] - 2022/06/04
 ### Added